+++ Runtime/Components/LocalizationTextBase.cs
@@ -1,79 +1,77 @@
+using System;
+using System.Collections.Generic;
 using SimplyLocalize.Runtime.Data;
 using SimplyLocalize.Runtime.Data.Keys.Generated;
 using SimplyLocalize.Runtime.Data.SearchableEnum;
 using SimplyLocalize.Runtime.Data.StringEnum;
 using SimplyLocalize.Runtime.Main;
 using UnityEngine;
-using UnityEngine.Events;
 
 namespace SimplyLocalize.Runtime.Components
 {
     public abstract class LocalizationTextBase : MonoBehaviour
     {
-        private static string _lang;
         private static FontHolder _overrideFontHolder;
+
+        [SearchableStringEnum, SerializeField] private StringEnum<LocalizationKey> _localizationKey;
         
-        [SearchableStringEnum]
-        [SerializeField] private StringEnum<LocalizationKey> _localizationKey;
-        
-        private bool _isInitialized;
-        
+        public bool IsInitialized { get; protected set; }
+
         public bool Translated { get; protected set; }
         public string DefaultText { get; protected set; }
         public LocalizationKey LocalizationKey => _localizationKey.Value;
+
+        public static FontHolder OverrideFontHolder => Localization.TryGetFontHolder(out _overrideFontHolder) ? _overrideFontHolder : null;
+
+        public static Dictionary<string, string> CurrentLocalization => Localization.CurrentLocalization;
         
-        private static LocalizationDictionary AllLocalizationsDict => Localization.LocalizationDictionary;
+        public event Action LanguageChanged;
 
         private void Awake()
         {
-            if (!_isInitialized) Init();
+            if (!IsInitialized)
+            {
+                Initialize();
+            }
+        }
+
+        private void OnDestroy()
+        {
+            Localization.LanguageChanged -= SetLanguage;
         }
 
         public void TranslateByKey(LocalizationKey key)
         {
             _localizationKey.Value = key;
 
-            if (_isInitialized == false) 
-                Init();
-
-            SetTextByKey();
+            if (!IsInitialized)
+            {
+                Initialize();
+            }
+            else
+            {
+                SetTextByKey();
+            }
         }
 
         public void TranslateByKey(string key)
         {
-            if (_isInitialized == false) 
-                Init();
-
-            SetTextByKey(key);
+            if (!IsInitialized)
+            {
+                Initialize();
+            }
+            else
+            {
+                SetTextByKey(key);
+            }
         }
 
-        public static bool TryGetKey(LocalizationKey localizationKey, out string key)
-        {
-            return LocalizationKeys.Keys.TryGetValue(localizationKey, out key);
-        }
-        
         /// <summary>
         /// Concatenates the given string 'str' with the 'DefaultText' and a separator.
         /// </summary>
         /// <param name="str">Additional string</param>
         /// <param name="separator">Separator between 'DefaultText' and 'str'</param>
         public abstract void Concatenate(string str, string separator = " ");
-
-        public static bool TryGetTranslatedText(LocalizationKey localizationKey, out string translated)
-        {
-            translated = "***";
-            return LocalizationKeys.Keys.TryGetValue(localizationKey, out var key) &&
-                   AllLocalizationsDict.TryGetTranslating(_lang, key, out translated);
-        }
-
-        public static void ApplyLocalization()
-        {
-            _lang = Localization.CurrentLanguage;
-            if (Localization.TryGetFontHolder(out var fontHolder))
-            {
-                _overrideFontHolder = fontHolder;
-            }
-        }
 
         protected abstract void SetTranslate(string translatedText);
         protected abstract void OnHasNotTranslated(string key);
@@ -83,19 +81,21 @@ namespace SimplyLocalize.Runtime.Components
             DefaultText = translatedText;
             Translated = true;
         }
-
-        protected abstract void SetFont(FontHolder overrideFontHolder);
         
+        protected abstract void SaveDefaultFont();
+        protected abstract void SetFont(FontHolder overrideFontHolder);
+        protected abstract void ResetFont();
+
         private void SetTextByKey()
         {
-            if (TryGetKey(LocalizationKey, out var key))
+            if (Localization.TryGetKey(LocalizationKey, out var key))
                 SetTextByKey(key);
             else Debug.LogWarning($"Localization key {LocalizationKey} not founded");
         }
 
         private void SetTextByKey(string key)
         {
-            if (AllLocalizationsDict.TryGetTranslating(_lang, key, out var translatedText))
+            if (CurrentLocalization.TryGetValue(key, out var translatedText))
             {
                 ApplyTranslate(translatedText);
                 SetTranslate(translatedText);
@@ -103,15 +103,30 @@ namespace SimplyLocalize.Runtime.Components
             else OnHasNotTranslated(key);
         }
 
-        private void Init()
+        private void Initialize()
+        {
+            SaveDefaultFont();
+            
+            SetLanguage();
+            Localization.LanguageChanged += SetLanguage;
+
+            IsInitialized = true;
+        }
+        
+        private void SetLanguage()
         {
             SetTextByKey();
 
-            if (_overrideFontHolder)
+            if (OverrideFontHolder)
             {
-                SetFont(_overrideFontHolder);
+                SetFont(OverrideFontHolder);
             }
-            _isInitialized = true;
+            else
+            {
+                ResetFont();
+            }
+            
+            LanguageChanged?.Invoke();
         }
     }
 }